Removed some duplication of code between ImportMapper classes
